Add new binding interface
New binding interface is a 2D overlay interface that can be embeded in to
application.
To enable it, create a file named "vive_role_bindings.cfg",
or copy from "Assets/HTC.UnityPlugin/ViveInputUtility/Scripts/ViveRole/BindingInterface/BindingConfigSample"
and put it in the project or build folder, the interface will pop up
whenever the "toggle_interface_key_code" defind in the config file is
pressed.
If you don't want to use the hot key to open the interface, use
HTC.UnityPlugin.Vive.ViveRoleBindingsHelper.EnableBindingInterface to
open the interface manually.