Make scroll delta tunable from the Editor
Made the hard-coded scroll delta in Update() a member of the class so it can be easily modified from the Editor. Large, far away objects need to have the delta tuned so they can be easily moved without having to scroll too much. The default value of 0.01f has been preserved.