Add autoScaleReticle setting
The autoScaleReticle variable allows the reticle to scale itself so it maintains a desired size in screen/pixel space (set by the sizeInPixels variable). This makes it possible to lase both near and far objects without the reticle getting too big or too small; instead it will always appear to be the same size on screen no matter what you're pointing at. This setting defaults to off so as not to change the current behavior. This is an implementation of issue #104.