Cache reset poses so effectTargets maintain position
If the effectTargets are moved after the application is running when OnColliderEventPressUp restores the positions from resetPoses back into the effectTargets they'll get restored back to their original position in the scene instead of their updated positions. Called CacheResetPoses() from OnColliderEventPressEnter to update the cache so the restore works properly.